--- .env
@@ -1,1 +1,0 @@
-API_BASE_URL = "";
