+++ index.html
@@ -22,7 +22,6 @@
     <!-- <a href="/sign-in">Sign in</a> -->
 
     <script src="node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-    <script type="module" src="src/conduit-app.js"></script>
     <script src="/app.js" type="module"></script>
   </body>
 </html>
