+++ src/constants/url.config.js
@@ -1,0 +1,1 @@
+export const API_BASE_URL = `https://conduit.productionready.io/api`;
